rename loadServiceConfig func to loadEnvironmentConfig
