+++ perfomance-tests/tests/scnenarios-example.js
@@ -1,12 +1,12 @@
-﻿import { loadOptions, loadServiceConfig } from "../lib/utils.js";
+﻿import { loadOptions, loadEnvironmentConfig } from "../lib/utils.js";
 import {textSummary } from 'https://jslib.k6.io/k6-summary/0.0.1/index.js';
 
 export let options = loadOptions('default-scenarios');
-export let serviceConfig = loadServiceConfig('b2c');
+export let serviceConfig = loadEnvironmentConfig('b2c');
 
 export function setup() { return serviceConfig; }
 
-export { default as example } from './example.js';
+export { default as exampleAsScenarioTest } from './example-as-scenario-test.js';
 
 export default function (data) {
